api_yamdb/requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.6
Django==2.2.16
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
idna==3.4
iniconfig==1.1.1
isort==5.10.1
mccabe==0.7.0
packaging==21.3
pluggy==0.13.1
psycopg2-binary>=2.8,<2.9
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.21.0
pytz==2022.6
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.12
gunicorn==20.0.4

asgiref==3.2.10
